--- deepdespeckling/sar2sar/train/.directory
@@ -1,4 +1,0 @@
-[Dolphin]
-Timestamp=2022,4,5,17,1,21
-Version=4
-ViewMode=1
