+++ deepdespeckling/test_deepdespeckling.py
@@ -1,10 +1,9 @@
-from deepdespeckling.merlin.inference.despeckling import despeckle_from_coordinates
+from deepdespeckling.despeckling import despeckle, despeckle_from_coordinates, despeckle_from_crop
 
-image_path = "/Users/hadrienmariaccia/Documents/Projects/deepdespeckling/img/entire"
-destination_directory = "/Users/hadrienmariaccia/Documents/Projects/deepdespeckling/img/entire"
-model_weights_path = "/Users/hadrienmariaccia/Documents/Projects/deepdespeckling/deepdespeckling/merlin/inference/saved_model/spotlight.pth"
+image_path = "/Users/hadrienmariaccia/Documents/Projects/deepdespeckling/img/entire/sar2sar_tests"
+destination_directory = "/Users/hadrienmariaccia/Documents/Projects/deepdespeckling/img/entire/sar2sar_tests"
 coordinates_dictionnary = {'x_start': 0,
                            'y_start': 0, 'x_end': 700, 'y_end': 700}
 
-despeckle_from_coordinates(image_path, coordinates_dictionnary, destination_directory,
-                           model_weights_path=model_weights_path)
+despeckle_from_crop(image_path, destination_directory,
+                    model_name="sar2sar")
