+++ deepdespeckling/utils/constants.py
@@ -1,6 +1,3 @@
-from scipy import special
-import numpy as np
-
 # Global maximum and minimum values obtained empirically and used to normalize SAR images
 M = 10.089038980848645
 m = -1.429329123112601
